db.py now has the database refactored. All functionality of the class dbhandler is now in the model classes. A function can be found under the class it returns
server.py refactored to use this new db.py setup.
handlers.py also editted to use the new setup. iteration to make a recursive dict out of the arguments of a
action is moved to db.py as a function of ActionArg

shellsensor and domostime editted to use names instead of ident where applicable