revamped module structure and module rpc structure
Modules now support multiple rpc's of the same type. This means that the database relation between a sensor and module has moved to sensor and modulerpc. Furthermore, a module gets a modulerpc key with the registration reply for each sensor it gets.
Modules are autoloaded. This enforces the directorystructure of $module/$module.py The $module.py is loaded and the start() function in this module is executed after importing the module. The stop() function is meant to be called at shutdown to clean up anything
that the module was using.